% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.r
\name{build_polygon_df}
\alias{build_polygon_df}
\title{Transform a SpatialPolygonsDataFrame object into a data}
\usage{
build_polygon_df(polygons = NULL)
}
\arguments{
\item{polygons}{SpatialPolygonsDataFrame object.}
}
\value{
Dataframe with three columns: x coordinate, y coordinate,
   and polygon id.
}
\description{
\code{build_polygon_df} takes a SpatialPolygonsDataFrame object and strips 
coordinates of points which define the polygons. Each point also carries 
it's associated polygon id.
}

