% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.r
\name{makePNG}
\alias{makePNG}
\title{Write a .PNG file from a ggplot2 object.}
\usage{
makePNG(plt, filename, ht = 6, wt = 6, ppi = 300)
}
\description{
\code{makePNG} writes a ggplot2 object into a .PNG file.
}
\details{
@param plt ggplot2 object.
 @param filename Character string. File name and path (.png extension
   automatically added).
 @param ht Integer. Height of image in inches.
 @param wt Integer. Width of image in inches.
 @param ppi Integer. Pixels per inch resolution.
 @return Writes a file with the name and location speficied.
}

