% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.r
\name{plot_raster_points}
\alias{plot_raster_points}
\title{Plot raster using ggplot.}
\usage{
plot_raster_points(data = NULL)
}
\description{
\code{plot_raster} creates a ggplot of a raster.
}
\details{
@param data Raster. A single layer raster to be plotted.
 @return Returns a ggplot object.
}

